# Fix performance regression on simple cases of indexing