This PR fixes some of the performance regression on Tensors going from 0.3 to 0.4.

It dispatches the operations to the old kernels that we have, whenever possible.
This means that we use `index_select` / `masked_select` / `index_fill_` or `masked_fill_` when the arguments for the indexing allow it.

Ideally, we want to put this code in ATen, but at first sight it seemed simpler to change it here.

Here are the results from a quick micro benchmark:

```python
# master

a = torch.rand(100, 100, 100)
%timeit a[[2, 1, 0 ,4]]
196 µs ± 42.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

%timeit a[[2, 1, 0 ,4]] = 5
409 µs ± 203 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

%timeit a[a > 0.5]
44.7 ms ± 5.76 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit a[a > 0.5] = 1
50.5 ms ± 9.23 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# this branch

a = torch.rand(100, 100, 100)
%timeit a[[2, 1, 0 ,4]]
17.5 µs ± 2.19 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit a[[2, 1, 0 ,4]] = 5
13.6 µs ± 1.6 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit a[a > 0.5]
7.54 ms ± 1.34 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)

%timeit a[a > 0.5] = 1
5.81 ms ± 1.64 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

We see roughly 5-10x speedup on those simple cases.