@fmassa 2018-04-20
Closing in favour of https://github.com/pytorch/pytorch/pull/6793